backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
python-multipart
google-generativeai
smtplib-ssl
email-validator
pymongo
motor
PyJWT